Move navigation-related elements to the 'nav' element group folder
