+++ angular/components/elements/element-group-nav/nav-bar/nav-bar.component.scss
@@ -1,8 +1,8 @@
 // nav-bar.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
 
 $nav-bar-height: calc-size(6, 3, 2);
