+++ angular/components/elements/element-group-nav/nav-item/nav-item.component.scss
@@ -1,8 +1,8 @@
 // nav-item.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
 
 nav-item {
     @include flex(center, center, row);
